implemented sensor skills, fixed conditional passive states, tweaked sp regen

+++ Maple2Storage/Types/Metadata/ItemMetadata.cs
@@ -132,7 +132,7 @@ public class ItemSkillMetadata
     [XmlElement(Order = 1)]
     public int SkillId;
     [XmlElement(Order = 2)]
-    public int SkillLevel;
+    public short SkillLevel;
 }
 
 [XmlType]
